Fixed satResponseTimeout const being assigned value too large

+++ Source/CommCtrlrArduino/CommCtrlrConfig.h
@@ -1,6 +1,7 @@
 #ifndef CommCtrlrConfig_h
 #define CommCtrlrConfig_h
 
+
 #define __WHITESTARBALLOON__
 #define printExpectings
 #define printVerifiedModemSerialMessages
@@ -56,7 +57,7 @@ const unsigned int satPowerOffMinimumTime = 2000;    //Probably 2000 millis for 
 const unsigned int maxTelemLenConst = 340;          //Maximum acceptable length of telemetry packet FROM EEPROM, set by Iridium MO max message size
 #define satNetworkNotAvailable = 255;
 const int LongMsgQueueLen = 20;             // Number of messages that can be in the queue to send out sat modem
-const unsigned int satResponseTimeout = 2UL * 60UL * 1000UL;       // (ms) Timeout used when waiting for response timeouts
+const unsigned long satResponseTimeout = 2UL * 60UL * 1000UL;       // (ms) Timeout used when waiting for response timeouts
 const unsigned int satSBDIXResponseLost = 30000;     // (ms) How much time to wait before assuming SBDIX command failed
 
 /*******************************
